feat: add menu selection for installing packages by chapter
The code now includes a menu that allows the user to select a chapter for installing packages. If the user selects "Todo" (All), it will install all packages from a specific URL. If a specific chapter is selected, it will install only the packages associated with that chapter from a different URL. This provides more flexibility in package installation based on chapter selection.